Build RN entry file based on webpack dependency tree.
Previously, the entry file that was passed to the RN packager only require()'d 'react-native'. Requiring a RN module that was not in the main 'react-native' dependency tree would break, as the module was not included in the RN bundle generated by the packager.

This patch builds the entry file dynamically based on the modules required by the app source, allowing for any module inside react-native/**/* to be require()'d.

Fixes #14

/cc @auser